requirements.txt
# Core application
streamlit
pandas
SQLAlchemy
pymssql
python-dotenv
plotly

# Testing
pytest
